move objectToXML to its own file
