+++ src/llmTools/index.ts
@@ -1,5 +1,4 @@
 import type { WebClient } from '@slack/web-api/dist/WebClient';
-import { XMLBuilder } from 'fast-xml-parser';
 import type { ChatCompletionTool } from 'openai/resources/chat';
 import { logger } from '../services/logger';
 
@@ -71,21 +70,3 @@ export const getToolImplementationsMap = ({
     return acc;
   }, {});
 };
-
-// --- Utility Functions ---
-
-/**
- * Convert an object to XML. XML is a best practice format for feeding into LLMs
- *
- * @param obj - The object to convert
- * @returns The XML string
- */
-// biome-ignore lint/suspicious/noExplicitAny: we literally want to convert any object to XML
-export function objectToXml(obj: any): string {
-  const builder = new XMLBuilder({
-    format: true,
-    indentBy: '  ',
-  });
-
-  return builder.build(obj);
-}
