+++ src/llmTools/objectToXML.ts
@@ -1,0 +1,18 @@
+/**
+ * Convert an object to XML. XML is a best practice format for feeding into LLMs
+ *
+ * @param obj - The object to convert
+ * @returns The XML string
+ */
+
+import { XMLBuilder } from 'fast-xml-parser';
+
+// biome-ignore lint/suspicious/noExplicitAny: we literally want to convert any object to XML
+export function objectToXml(obj: any): string {
+  const builder = new XMLBuilder({
+    format: true,
+    indentBy: '  ',
+  });
+
+  return builder.build(obj);
+}
